Lecture8/requirements.txt
amqp==2.6.1
apns2==0.7.2
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.4.1
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.2
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
DAWG-Python==0.7.2
defusedxml==0.5.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==2.2
django-celery-beat==1.1.1
django-ckeditor==6.1.0
django-configurations==2.2
django-cors-headers==2.4.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-import-export==1.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-ordered-model==3.1.1
django-push-notifications==2.0.0
django-rest-framework==0.1.0
django-reversion==3.0.4
django-suit==0.2.28
django-suit-locale==1.0.10
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.15.0
docopt==0.6.2
ephem==3.7.6.0
et-xmlfile==1.0.1
eventlet==0.24.1
Fabric3==1.14.post1
gevent==1.3.5
greenlet==0.4.14
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
importlib-metadata==3.7.2
ipaddr==2.2.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jinja2schema==0.1.4
kombu==4.6.11
lockfile==0.12.2
lxml==4.4.0
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
numpy==1.21.2
odfpy==1.4.0
olefile==0.45.1
openpyxl==2.5.12
packaging==20.3
pandas==1.3.3
paramiko==2.8.0
pep517==0.8.2
phonenumbers==8.9.10
pika==0.12.0
Pillow==8.4.0
progress==1.5
prompt-toolkit==3.0.22
psycopg2==2.8.6
psycopg2-binary==2.8.6
py-vapid==1.8.2
pyasn1==0.4.4
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.2
PyNaCl==1.4.0
pyotp==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.2
python-dotenv==0.19.1
python-telegram-bot==13.8.1
pytoml==0.1.21
pytz==2021.3
pywebpush==1.13.0
PyYAML==3.13
requests==2.26.0
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.2.10
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==3.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
